Add Decision Panel
Add a decision panel to be used for making decisions in the game.

Also fix a few battle-loading bugs related to the wrong prefab stats
being loaded into battle by changing the name of the enemy prefabs.

Also fix a bug that was causing defeating Tentacle Monster to remove
Boss Spawn.